requirements.txt
aiogram==2.11.2
emoji==0.6.0
Pillow==8.0.1
pytz==2020.5
